IGNITE-17736 Ducktape test to measure consistency repair performance
